// src/app/Models/testSection-Tech.ts
export class TestSectionTech {
  id?: number;
  testsectionName?: string;
  experience?: string;
  qtsNumber?: number;
  difficulty?: string;
  difficulties?: string[];
  publicNumber?: number[];
  privateNumber?: number[];
  domain_id?: number[];
  createdAt?: Date;
  userUUID?: string;
  uuid?: string;
}